+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Allow restriction pagination types globally, for a schema or by passing an
+  option.
 - Allow global configuration of `get_cursor_value_func`, `max_limit` and
   `default_limit`.
 - Add `Flop.option` type, improve documentation of available options.